# LATCH_GO_PORT_API_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Latch API",
    "version": "1.0.0",
    "description": "Machine-readable spec for porting Latch Next.js /api/* routes to Go. Extension-verified. See LATCH_GO_PORT_GUIDE.md."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Go dev server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Next.js reference server"
    }
  ],
...
  "paths": {
...
    "/api/multisig/drafts/{id}": {
      "get": {
        "summary": "GET /api/multisig/drafts/{id}",
        "operationId": "_api_multisig_drafts_id_get",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/drafts/[id]/route.ts",
        "x-latch-auth": "session+creator",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL"
        ],
        "x-latch-db-models": [
          "MultisigDraft"
        ],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "draft": {
                      "$ref": "#/components/schemas/SerializedDraft"
                    },
                    "inviteUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "draft": {
                      "$ref": "#/components/schemas/SerializedDraft"
                    },
                    "inviteUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "summary": "PATCH /api/multisig/drafts/{id}",
        "operationId": "_api_multisig_drafts_id_patch",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/drafts/[id]/route.ts",
        "x-latch-auth": "session+creator",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL"
        ],
        "x-latch-db-models": [
          "MultisigDraft"
        ],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/multisig/drafts/{id}/webauthn/register/begin": {
      "post": {
        "summary": "POST /api/multisig/drafts/{id}/webauthn/register/begin",
        "operationId": "_api_multisig_drafts_id_webauthn_register_begin_post",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/drafts/[id]/webauthn/register/begin/route.ts",
        "x-latch-auth": "session+creator",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL",
          "WEBAUTHN_RP_ID",
          "WEBAUTHN_ORIGIN"
        ],
        "x-latch-db-models": [],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/multisig/drafts/{id}/webauthn/register/finish": {
      "post": {
        "summary": "POST /api/multisig/drafts/{id}/webauthn/register/finish",
        "operationId": "_api_multisig_drafts_id_webauthn_register_finish_post",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/drafts/[id]/webauthn/register/finish/route.ts",
        "x-latch-auth": "session+creator",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL",
          "WEBAUTHN_RP_ID",
          "WEBAUTHN_ORIGIN"
        ],
        "x-latch-db-models": [],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/multisig/drafts/{id}/webauthn/authenticate/begin": {
      "post": {
        "summary": "POST /api/multisig/drafts/{id}/webauthn/authenticate/begin",
        "operationId": "_api_multisig_drafts_id_webauthn_authenticate_begin_post",
        "tags": [
          "multisig"
        ],
        "x-latch-source-file": "app/api/multisig/drafts/[id]/webauthn/authenticate/begin/route.ts",
        "x-latch-auth": "session+creator",
        "x-latch-env-required": [
          "DATABASE_URL",
          "DIRECT_URL",
          "WEBAUTHN_RP_ID",
          "WEBAUTHN_ORIGIN"
        ],
        "x-latch-db-models": [],
        "x-latch-legacy-error": true,
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid",
        "description": "30-day sliding session. Auto-created on first request."
      }
    },
    "schemas": {
...
      "LegacyErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "SerializedDraft": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "threshold": {
            "type": "integer"
          },
          "accountSaltHex": {
            "type": "string"
          },
          "inviteToken": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "collecting",
              "deployed"
            ]
          },
          "predictedAddress": {
            "type": [
              "string",
              "null"
            ]
          },
          "smartAccountAddress": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "integer"
          },
          "expiresAt": {
            "type": [
              "integer",
              "null"
            ]
          },
          "validMemberCount": {
            "type": "integer"
          },
          "canDeploy": {
            "type": "boolean"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  },
...
}