# app/api/multisig/drafts/[id]/webauthn/authenticate/begin/route.ts
import { NextResponse } from "next/server";
import { getOrCreateSession } from "@/lib/session";
import { getDraftForCreator } from "@/lib/multisig-draft";
import { beginDraftWebauthnAuthentication } from "@/lib/multisig-draft-webauthn";
import { WebauthnCeremonyConfigError } from "@/lib/webauthn-server";

export const runtime = "nodejs";

type Ctx = { params: Promise<{ id: string }> };

export async function POST(request: Request, ctx: Ctx) {
  try {
    const { userId } = await getOrCreateSession();
    const { id } = await ctx.params;
    const draft = await getDraftForCreator(id, userId);
    if (!draft || draft.status !== "collecting") {
      return NextResponse.json({ error: "Draft not found" }, { status: 404 });
    }

    const body = (await request.json().catch(() => ({}))) as { chromeExtensionId?: string };
    const { options } = await beginDraftWebauthnAuthentication({
      request,
      draftId: id,
      challengeUserId: userId,
      chromeExtensionId: body.chromeExtensionId,
      sessionUserId: userId,
    });

    return NextResponse.json({ options });
  } catch (error) {
    if (error instanceof WebauthnCeremonyConfigError) {
      return NextResponse.json({ error: error.message }, { status: 400 });
    }
    return NextResponse.json(
      { error: error instanceof Error ? error.message : "Authentication begin failed" },
      { status: 500 }
    );
  }
}